lab_02/requirements.txt
pyspark==3.5.0
psycopg2-binary==2.9.9
clickhouse-driver==0.2.6
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.26.2 